backend/forensic_engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
qdrant-client==1.7.0
pydantic==2.5.3
python-dotenv==1.0.1
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
requests>=2.31.0
